Split DrawRoutes + related code into new module
